+++ src/elements/map-control-mouse-position/config.js
@@ -1,0 +1,3 @@
+export const elementName = 'map-control-mouse-position';
+export const defaultViewProjection = 'EPSG:4326';
+export const defaultCoordinateDecimals = 4;