+++ src/web-gis-elements-lite.js
@@ -14,6 +14,7 @@ import './elements/map-layer-geojson';
 import './elements/map-layer-singlepoint';
 
 import './elements/map-control-defaults';
+import './elements/map-control-mouse-position';
 import './elements/map-control-simple-layer-list';
 import './elements/map-control-scale-line';